Fixed offsets and sigs + how-to in commit descr.
Seems like crosshair ID changes often.
To find it, look up `C_CSPlayer::GetIDTarget(void)` in Mac binaries. The *mov* right above the first *test* instruction contains our offset - 0xB398.

We can use `CountdownTimer::Now(void)` to find it on Linux. The function below contains "m_timestamp" string, let's use this as starting point. Look up its first reference on Linux. Above that whole subroutine is another one with "IntervalTimer" and above that is the `CountdownTimer::Now(void)` function we need (looks pretty similiar to the one in Mac binaries).

Now find all references to that function (Ctrl+X with the cursor on sub_xxxxx). We should see ~60 lines. On macOS, we can see that the function we're interested in - `C_CSPlayer::GetIDTarget(void)` - is on lines 27 and 28, 0x28 bytes apart. In Linux binaries, we can see the same pattern. The function is, however, on lines 25 and 26 and 0x37 bytes apart. Close enough. Doubleclick the line 25.

The function we just found is `C_CSPlayer::GetIDTarget(void)`. As I said, the first *mov* above the first *test* instruction contains our offset.

I know, that this is a pretty ghetto way of finding it, but that's how I did it.